Fix an issue with small tap area of the ellipsis in the post list
